Use threads for UnixAgent clients
OpenSSH since v6.8 or so keeps the socket to the agent open as long as the
ssh client program is running. This means that only one client was able
to be serviced at a time. This uses threads to be able to service more than
one client at a time.

Issue dlech/KeeAgent#142