Update nuget packages and VS version
Update nuget packages to latest versions. This includes updating
BouncyCastle to 1.8.1, updating NUnit tests to 3.6.0, adding a Mono.Posix
package so we can build on Windows, and adding a Chaos.NaCl package instead
of using a git submodule and project reference.

Also updated to VS 2017 RC, which automatically made some changes as well.