PageantAgent: ensure background thread is stopped on Dispose
This joins the background thread when PageantAgent is disposed to
ensure that system resources have been freed by the type Dispose()
returns.

Also add NonParallelizable attribute to tests that create a PageantAgent
to ensure they don't fail because they are trying to use the same
system resources at the same time.