If Windows OpenSSH agent support is enabled and subsequently disabled, the agent named pipe will not be closed. This is a (minor) security vulnerability as KeeAgent will continue to serve SSH agent responses when the functionality is meant to be disabled.
There's two reasons why this happens:
- While Dispose() is called on the class, a new server is started by
  listenerThread(). Also, thread objects aren't garbage collected, so
  this won't eventually fix itself when the garbage collector runs.
- The call to WaitForConnection() is a blocking call. That extends to
  the Dispose() call, which will be blocked until a connection to the
  pipe is made. It'll then be immediately closed, so the client won't
  likely get any response, but per above a new named pipe is created.

A side-effect of the latter is that re-enabling the feature in the same
KeePass process will never work, as the named pipe path will always be
still in use. An error message to that effect is shown to the user.

Switching to asynchronous connections helps solve both of these issues.
Instead of spawning threads we leverage the .NET thread pool, and the
Dispose() logic has been improved to ensure the listening server will be
reliably closed if the feature is disabled.

Suggested-by: @ralish