implement progress callbacks for decryption
This adds progress callbacks for decryption functions. Keys with key derivation functions (bcrypt in openssh keys and argon2 in putty keys) can take a long time, so we need progress feedback to show the user.

The bcrypt algorithm is modified to provide progress. However the argon2 implementation is 3rd party and doesn't have the required hooks.