Catch and ignore failures to retrieve client PID for OpenSSH agent
There's an inherent race condition where the process on the other end of
the named pipe may be gone by the time we attempt to retrieve a Process
object for it. I don't know the exact circumstances this happens in, but
it seems to be common when using WSL. Without the catch, the exception
is unhandled and effectively kills the named pipe server.