Drop mutable public key from private key.
We were allowing to "override" the public key for a SshPrivateKey object.
However, this lead to problems with usage of knowing when to override
the value, or forgetting to in some cases. So instead we make the
private key object immutable and only provide the public key if it is
included in the private key file.